requirements.txt
streamlit
python-dotenv
python-docx
PyPDF2
google-generativeai
